+++ package.json
@@ -14,6 +14,7 @@
     "@mui/icons-material": "^6.4.4",
     "@mui/material": "^6.4.4",
     "@mui/x-date-pickers": "^7.27.1",
+    "@supabase/ssr": "^0.5.2",
     "dayjs": "^1.11.13",
     "next": "15.1.7",
     "react": "^19.0.0",
